Add support for KaTeX math through the goldmark-katex extension.
This requires a little bit of refactoring of static asset serving,
because instead a very limited number of assets we now need to serve
the complete KaTeX support package consisting of CSS and fonts.

The CSS and fonts are taken from KaTeX version 0.16.4. We are only
shipping the .woff2 versiions, because we need a modern browser
anyways and can assume support.

The rendering is done on server side, so we don't need to include
Javascripts. A browser with MathML support is needed, Firefox had
support since forever but Chromium got tit only recently in 109.